ScriptHookVDot3移行, 非同期処理をasync/awaitベースに修正, 全MODの挙動を見直し, LemonUI対応

+++ Inferno/InfernoScripts/Parupunte/Scripts/MultiLeggedRace.cs
@@ -1,31 +1,31 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using GTA;
 using GTA.Math;
 using GTA.Native;
-using UniRx;
+using Inferno.Utilities;
 
 namespace Inferno.InfernoScripts.Parupunte.Scripts
 {
     [ParupunteConfigAttribute("参　勤　交　代", "大　政　奉　還")]
     [ParupunteIsono("さんきんこうたい")]
-    class MultiLeggedRace : ParupunteScript
+    internal class MultiLeggedRace : ParupunteScript
     {
+        private List<Ped> _allPedList = new();
+
         public MultiLeggedRace(ParupunteCore core, ParupunteConfigElement element) : base(core, element)
         {
         }
 
-        private List<Ped> explosionList = new List<Ped>();
-        private List<Ped> allPedList = new List<Ped>();
-
         public override void OnStart()
         {
-
             ReduceCounter = new ReduceCounter(15 * 500);
             AddProgressBar(ReduceCounter);
 
-            this.OnUpdateAsObservable
+            OnUpdateAsObservable
                 .Subscribe(_ =>
                 {
                     if (core.PlayerPed.IsDead)
@@ -36,11 +36,15 @@ namespace Inferno.InfernoScripts.Parupunte.Scripts
 
             ReduceCounter.OnFinishedAsync.Subscribe(_ =>
             {
-                foreach (var ped in allPedList)
+                foreach (var ped in _allPedList)
                 {
-                    GTA.World.AddExplosion(ped.Position, GTA.ExplosionType.Grenade, 1.0f, 0.0f);
-                    ped.Kill();
+                    if (ped.IsSafeExist())
+                    {
+                        GTA.World.AddExplosion(ped.Position, GTA.ExplosionType.Grenade, 1.0f, 0.0f);
+                        ped.Kill();
+                    }
                 }
+
                 ParupunteEnd();
             });
 
@@ -49,51 +53,43 @@ namespace Inferno.InfernoScripts.Parupunte.Scripts
             {
                 Function.Call(Hash.REQUEST_NAMED_PTFX_ASSET, ptfxName);
             }
-            Function.Call(Hash._SET_PTFX_ASSET_NEXT_CALL, ptfxName);
 
-            allPedList =
-            core.CachedPeds.Where(x => x.IsSafeExist() && !x.IsRequiredForMission() && !x.IsCutsceneOnlyPed())
-                .Take(30).ToList();
+            Function.Call(Hash.USE_PARTICLE_FX_ASSET, ptfxName);
+
+            _allPedList =
+                core.CachedPeds.Where(x =>
+                        x.IsSafeExist() && !x.IsRequiredForMission() && !x.IsCutsceneOnlyPed() && x.IsAlive)
+                    .Take(50)
+                    .ToList();
 
             var playerPos = core.PlayerPed.Position;
-            var num = 5;
+            var num = 10;
 
-            var i = -allPedList.Count / 2;
-            foreach (var p in allPedList.Where(x=>x.IsSafeExist()))
+            var i = -_allPedList.Count / 2;
+            foreach (var p in _allPedList.Where(x => x.IsSafeExist()))
             {
                 p.Task.ClearAllImmediately();
                 p.Position =
                     playerPos
-                    + core.PlayerPed.RightVector * 1.5f * (i / num)
+                    + core.PlayerPed.RightVector * 2.5f * (i / num)
                     + core.PlayerPed.ForwardVector * (Math.Abs(i) % num)
-                    + core.PlayerPed.ForwardVector;
+                    + core.PlayerPed.ForwardVector * 5;
 
                 if (p.IsSafeExist())
                 {
                     AutoReleaseOnParupunteEnd(p);
-                    p.Rotation = core.PlayerPed.Rotation;
-                    p.Health = 1;
-                    StartCoroutine(DashCoroutine(p));
+                    p.Rotation = Quaternion.RotationAxis(core.PlayerPed.UpVector, (float)Math.PI) *
+                                 core.PlayerPed.Rotation;
+                    p.Health = 100;
+                    DashAsync(p, ActiveCancellationToken).Forget();
                     i++;
                 }
-
             }
-
-
         }
 
-
-        protected override void OnUpdate()
+        protected override void OnFinished()
         {
-            base.OnUpdate();
-            var targetList = explosionList.Take(1).ToArray();
-            foreach (var ped in targetList)
-            {
-                GTA.World.AddExplosion(ped.Position, GTA.ExplosionType.Grenade, 1.0f, 0.2f);
-                ped.Kill();
-                explosionList.Remove(ped);
-            }
-
+            _allPedList.Clear();
         }
 
         private void SetAnimRate(Ped ped, float rate)
@@ -101,17 +97,21 @@ namespace Inferno.InfernoScripts.Parupunte.Scripts
             Function.Call(Hash.SET_ANIM_RATE, ped, (double)rate, 0.0, 0.0);
         }
 
-        private IEnumerable<object> DashCoroutine(Ped ped)
+        private async ValueTask DashAsync(Ped ped, CancellationToken ct)
         {
             var speed = 5;
             var isReturedCount = 10;
-            while (!ReduceCounter.IsCompleted)
+            while (!ReduceCounter.IsCompleted && !ct.IsCancellationRequested)
             {
-                if (!ped.IsSafeExist()) yield break;
+                if (!ped.IsSafeExist())
+                {
+                    return;
+                }
+
                 if (ped.IsDead)
                 {
-                    explosionList.Add(ped);
-                    yield break;
+                    GTA.World.AddExplosion(ped.Position, GTA.ExplosionType.Grenade, 1.0f, 0.2f);
+                    return;
                 }
 
                 if (isReturedCount == 0)
@@ -119,8 +119,11 @@ namespace Inferno.InfernoScripts.Parupunte.Scripts
                     isReturedCount = 10;
                     ped.Quaternion = Quaternion.RotationAxis(ped.UpVector, (float)Math.PI) * ped.Quaternion;
                 }
-
-                if (isReturedCount > 0) isReturedCount--;
+                
+                if (isReturedCount > 0)
+                {
+                    isReturedCount--;
+                }
 
                 SetAnimRate(ped, speed);
                 Function.Call(Hash.SET_OBJECT_PHYSICS_PARAMS, ped, 200000000.0, 1, 1000, 1, 0, 0, 0, 0, 0,
@@ -133,7 +136,7 @@ namespace Inferno.InfernoScripts.Parupunte.Scripts
                     0, 0, 0);
                 StartFire(ped);
 
-                yield return null;
+                await Delay100MsAsync(ct);
             }
         }
 
@@ -143,14 +146,14 @@ namespace Inferno.InfernoScripts.Parupunte.Scripts
             var rotation = new Vector3(0.0f, 0.0f, 0.0f);
             var scale = 2.0f;
 
-            Function.Call(Hash._SET_PTFX_ASSET_NEXT_CALL, "core");
+            Function.Call(Hash.USE_PARTICLE_FX_ASSET, "core");
             Function.Call<int>(Hash.START_PARTICLE_FX_NON_LOOPED_ON_PED_BONE, "ent_sht_electrical_box",
-                ped, offset.X, offset.Y, offset.Z, rotation.X, rotation.Y, rotation.Z, (int)Bone.SKEL_L_Toe0, scale,
+                ped, offset.X, offset.Y, offset.Z, rotation.X, rotation.Y, rotation.Z, (int)Bone.SkelLeftToe0, scale,
                 0, 0, 0);
 
-            Function.Call(Hash._SET_PTFX_ASSET_NEXT_CALL, "core");
+            Function.Call(Hash.USE_PARTICLE_FX_ASSET, "core");
             Function.Call<int>(Hash.START_PARTICLE_FX_NON_LOOPED_ON_PED_BONE, "ent_dst_elec_fire",
-                ped, offset.X, offset.Y, offset.Z, rotation.X, rotation.Y, rotation.Z, (int)Bone.SKEL_R_Toe0, scale,
+                ped, offset.X, offset.Y, offset.Z, rotation.X, rotation.Y, rotation.Z, (int)Bone.SkelRightToe0, scale,
                 0, 0, 0);
         }
     }
